Get and set site data in object cache
Look for our site object data in a custom cache key. It seems like we could use the core functionality for this, but core really isn't multiple site ID friendly, so we might as well handle the just in case until I better understand what wpmu_current_site() can do for this.

Closes #1